ai_service/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
fastapi==0.116.1
h11==0.16.0
httptools==0.6.4
idna==3.10
lxml==6.0.1
orjson==3.10.7
pillow==10.4.0
pydantic==2.11.7
pydantic_core==2.33.2
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
python-docx==1.1.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
